binance-algo-trading-bot/requirements.txt
python-binance
TA-Lib
numpy
websocket_client
numpy
backtrader
flask